whitespace handling and pattern return (#35)
# Description

This addresses https://github.com/sbdchd/codeowners/issues/34 and also updates `matching_line` method to return the path alongside the owners. This is very helpful for others who may want to parse the CODEOWNERS file but also need access to the path pattern.

Also updates `.gitignore` to ignore pycache

I am unsure how you want to handle the versioning for this, I imagine it should bump the minor version but will differ to you. Would you like me to do it in this PR or would you do it yourself?

# Testing

- [x] CI