Support gitlab sections (#38)
Hi!

This pull request is seeking to introduce support for determining the GitLab section name of a file based on a CODEOWNERS file. (When section names are used)

https://docs.gitlab.com/ee/user/project/code_owners.html#code-owners-sections

This builds on https://github.com/sbdchd/codeowners/pull/23 to ensure folks can easily get the section name in addition to determining if a file is owned based on CODEOWNERS without just ignoring the section name information.

For folks managing a large monorepo with CODEOWNERS on GitLab, the sections are a nice way to group code by "system name."   Exposing this information via the codeowners library is really valuable for folks that need to determine how much code/data is associated with each system living in that monorepo as well as how much code is owned vs. unowned.

Thanks!

Brian